requirements.txt
pillow-simd>=9.0.0.post1
ujson>=5.3.0
numpy>=1.22.4
nb-cli >=0.6.0
nonebot-adapter-onebot>=2.1.0
fake_useragent>=0.1.11
pyttsx3>=2.71
pypinyin>=0.46.0
pypinyin_dict>=0.2.0